lobby search (#200)
* Autocompleted search for Lobby

* Update ListView to show filtered lobbies

* Disposing old binding

In order to prevent a potential memory leak, the old
AutoCompletionBinding is disposed. As this is done in a lambda
expression, a reference must be stored to dispose the old one. This is
done using an AtomicReference. If this is a suitable approach is open
for discussion.
In order to prevent simultaneous changes of the binding reference, the
whole action is synchronized.

* Filter lobby list in real time

* Replacing synchronized with 'getAndSet(...)'

* Removing Auto-completion feature for lobby search

As the search does immediately update the lobby list, there is no real
need to suggest auto-completions. Also it looks cleaner if there is no
comboBox over the lobby list thus blocking the really interesting part.