Resolving log error in docker container (#250)
Since the root folder of the server application did still belong to the
root, the kurtama user was not allowed to do changes within the
directory. This has been fixed by making the kurtama user and group the
owner of the directory.

Furthermore, the application folder has been remaned from /opt/app to
/opt/kurtama for clarity sakes.

A docker-compose.yml file was created for easier interacting with the
container.